refactor: Restructure Android project with package organization
