+++ mobile/app/src/main/AndroidManifest.xml
@@ -11,16 +11,16 @@
         android:theme="@style/Theme.FleetForge"
         tools:targetApi="31">
         <activity
-            android:name=".RegistrationActivity"
+            android:name=".activities.auth.RegistrationActivity"
             android:exported="false" />
         <activity
-            android:name=".ResetPasswordActivity"
+            android:name=".activities.auth.ResetPasswordActivity"
             android:exported="false" />
         <activity
-            android:name=".ForgotPasswordActivity"
+            android:name=".activities.auth.ForgotPasswordActivity"
             android:exported="false" />
         <activity
-            android:name=".LoginActivity"
+            android:name=".activities.auth.LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
