+++ mobile/app/src/main/java/com/ognjen/fleetforge/activities/auth/RegistrationActivity.java
@@ -1,8 +1,11 @@
-package com.ognjen.fleetforge;
+package com.ognjen.fleetforge.activities.auth;
 
 import android.os.Bundle;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.ognjen.fleetforge.R;
+import com.ognjen.fleetforge.fragments.auth.RegisterContactFragment;
 
 public class RegistrationActivity extends AppCompatActivity {
