+++ mobile/app/src/main/java/com/ognjen/fleetforge/activities/auth/ResetPasswordActivity.java
@@ -1,4 +1,4 @@
-package com.ognjen.fleetforge;
+package com.ognjen.fleetforge.activities.auth;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -8,6 +8,7 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import com.google.android.material.button.MaterialButton;
 import com.google.android.material.textfield.TextInputEditText;
+import com.ognjen.fleetforge.R;
 
 public class ResetPasswordActivity extends AppCompatActivity {
