+++ mobile/app/src/main/java/com/ognjen/fleetforge/fragments/common/PasswordChangeProfile.java
@@ -1,4 +1,4 @@
-package com.ognjen.fleetforge;
+package com.ognjen.fleetforge.fragments.common;
 
 import android.os.Bundle;
 
@@ -9,6 +9,9 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 
+import com.ognjen.fleetforge.R;
+import com.ognjen.fleetforge.fragments.driver.DriverProfile;
+import com.ognjen.fleetforge.fragments.user.UserProfile;
 import com.ognjen.fleetforge.model.UserRole;
 import com.ognjen.fleetforge.utils.AuthManager;
