+++ mobile/app/src/main/res/layout/fragment_password_change_profile.xml
@@ -4,7 +4,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PasswordChangeProfile">
+    tools:context=".fragments.common.PasswordChangeProfile">
 
     <Button
         android:id="@+id/changeBtn"
