+++ mobile/app/src/main/res/layout/fragment_register_profile.xml
@@ -52,7 +52,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:scaleType="centerCrop"
-                android:src="@drawable/profile"
+                android:src="@drawable/ic_profile"
                 android:backgroundTint="#E0E0E0"
                 app:shapeAppearance="@style/ShapeAppearance.Material3.Corner.Full" />
